Allow for archiving to /dev/null
Current workflow is that you must specify a destination for the
archival service. But what if you just want the bot to clear the
threads?

With this update, if you specify “/dev/null”, “None”, or “Nowhere”, or
lowercase variants, the bot will just remove the threads from the page
and not post them to an archive.

This code has not been tested and I cannot guarantee it will work.